Reorganize signup-related components and update import paths
