+++ src/signup/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../context/AuthContext"; // adjust path if needed
+import { useAuth } from "./AuthContext"; // adjust path if needed
 
 export default function ProtectedRoute({ children }) {
   const { isAuthenticated } = useAuth();
